Change default reconnect behavior
By default we should limit the number of times we attempt to reconnect
to Graphite as well as the number of times we try to send data. By
default we limit to three attempts. If a number less than or equal to
zero is specified in the config file we will retry forever.