Consistent usage of cfg module as a global namespace
Previously it was passed around in some cases (*Server) and used
as a global in others (udpserver, names).
While I see value in being able to re-use bucky modules separately,
it clearly was broken already (also by the late aggregation-db
commit).
Not sure if it's worth fixing, but it can certainly be done later.