Defer joining hostname and statname and allow custom clients.
Not all potential clients need hostnames and metric names joined together, so
we defer calling names.statname() until just before shipping data to Carbon.
Messages in the queue are now 4-tuples.

Also added a simple hook for custom clients. A client must respond to init/1
and send/4. They're not run in separate threads, so some care should be taken
to not hog interpreter time.

Updated tests to pass with the new message format.