remove log.debug statemens from udpserver when not running in debug mode and use arguments for string formatting on log functions
the log.debug statements are taking up a significant part of execution time (as measured by cProfile) this patch makes it so that these statement are only loaded when debug mode is True.

Changing all places (where possible) which call logger functions in combination with string formatting to use arguments on the logger function instead of doing the string formatting in place. This is a (micro-)optimalisation when the message will not be logged