Add a simple processing hook
A Processor is a process that recieves samples as they are parsed
by the servers and performs actions on them before handing them
over to the clients.

If a callable is defined in the 'processor' configuration variable,
a Processor process will aply this callable to the sample recieved
(host, name, val, time) and expects back a tuple of the same
structure to forward to clients, or None to drop the sample.