Make Bucky die gracefully
Add shutdown() method to Bucky that:
 - Calls close() on each server (sends b'EXIT' to the socket)
 - Puts a None in each client's pipe
 - Joins all servers and clients
 - If any other children remain alive, terminates them and raises
   exception.

Change SIGTERM handler to simply put a None in the queue read
by the parent process (Bucky) so that it calls its shutdown()
method and exits cleanly. No need to be putting multiple None's
in the queue, the parent process is the only one reading them.