Handle collectd counter/derive conflict
If a collectd metric is received with a value of type counter when
our types.db define it as derive, or vice versa, don't raise an
exception and assume the server's types.db is correct.

Types counter and derive are very similar. Also, it's common
for different versions/installations of collectd in 'clients'
to have a bit different definitions for the same metrics
(counter/derive conflict). This commit makes bucky's collectd
server handle such cases gracefully by assuming the server's
types.db is correct in case of conflict.